vim: make session management behave the way I like
The session is created manually - it's still possible to do simple
editions without using session management.
On opening the editor it notifies if a session is found in the current
directory.
The session is loaded manually.
If the session was loaded manually, the editor will autosave it on exit;
it won't touch the session file if it exists but was not loaded.